% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cleandns}
\alias{cleandns}
\title{Tranform a vector of DNS entries to a clean data frame}
\usage{
cleandns(dnames, strip.len = 6)
}
\arguments{
\item{dnames}{the input vector of names}

\item{strip.len}{numeric on which to remove variables}
}
\description{
given a vector of DNS entries, this will call tldextract and do
the following clean up tasks:
* remove any invalid entries (no top level domain matched)
* remove any duplciated entries
* strip any domain less than \code{strip.len} characters long.
}

