% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAfunctions.R
\name{CompLogML}
\alias{CompLogML}
\title{Computes Marginal Likelihoods for Each Clique and Value of Nmissing}
\usage{
CompLogML(D, Nmissing, delta)
}
\arguments{
\item{D}{A marginal table of the list overlap counts.}

\item{Nmissing}{The vector of possible values for the missing cell.}

\item{delta}{The prior hyper parameter for the Dirichlet distribution. This
has been properly re-scaled in bma.cr so that all marginal tables have
consistent priors.}
}
\value{
The log marginal likelihood of the marginal table.
}
\description{
Assembles all of the pieces of the marginal likelihoods to be used to
calculate the posterior probability of each model/value of Nmissing.
}
\examples{

## The function is currently defined as
function (D, delta)
{
    out <- apply(lgamma(D + delta), 1, sum) - apply(lgamma(D *
        0 + delta), 1, sum)
    return(out)
  }

}
\references{
Madigan, David, and Jeremy C. York. "Bayesian methods for
estimation of the size of a closed population."  Biometrika 84.1 (1997):
19-31.
}
\author{
James Johndrow \email{james.johndrow@gmail.com} and Kristian Lum
\email{kl@hrdag.org}
}
\keyword{BMA}
\keyword{capture-recapture}
\keyword{estimation}
\keyword{multiple}
\keyword{systems}
