% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgaFast.R
\docType{package}
\name{dgaFast}
\alias{dgaFast}
\alias{dgaFast-package}
\title{dgaFast: Multiple systems estimation using decomposable graphical models.}
\description{
This is an efficient re-implementation of the \code{dga} package of James Johndrow, Kristian Lum and Patrick Ball (2015) which implements Madigan and York (1997).
}
\author{
\strong{Maintainer}: Olivier Binette \email{olivier.binette@gmail.com}

Authors:
\itemize{
  \item Rebecca Steorts
}

}
