% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dgaPredict}
\alias{dgaPredict}
\title{Classify a vector of DNS names}
\usage{
dgaPredict(dns, dgaThreshold = 0.5)
}
\arguments{
\item{dns}{vector of dns names to classify}

\item{dgaThreshold}{the threshold at which to split DGA from legit}
}
\description{
Given one or more DNS names, this will predict whether or not it is generated
by an algorithm or if it is a legitamate domain name.
}

