% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dga.R
\name{dgaPredictDomain}
\alias{dgaPredictDomain}
\title{Classify a vector of stripped domain names}
\usage{
dgaPredictDomain(dns, dgaThreshold = 0.5)
}
\arguments{
\item{dns}{vector of stripped domain names to classify}

\item{dgaThreshold}{the threshold at which to split DGA from legit}
}
\description{
Given one or more stripped domain names (just the domain portion of the name), 
this will predict whether or not it is generated
by an algorithm or if it is a legitamate domain name.
}

