% Generated by roxygen2 (4.0.1): do not edit by hand
\name{entropy}
\alias{entropy}
\title{Calculate the entropy for a given string}
\usage{
entropy(instr)
}
\arguments{
\item{instr}{character or vector of characters to calculate entropy on}
}
\description{
Given a string this will return a numric with the entropy of the
characters in the string.
}

