% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dotaMatchesFunctions.R
\name{GetMatchHistoryAll}
\alias{GetMatchHistoryAll}
\title{Get history of all matches with 32bit Steam ID}
\usage{
GetMatchHistoryAll(accountId, apiKey)
}
\arguments{
\item{accountId}{32bit Steam ID}

\item{apiKey}{A api key string from https://steamcommunity.com/dev/apikey}
}
\value{
data.frame of match records
}
\description{
Get history of all matches with 32bit Steam ID
}

