% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dga.R
\name{getngram}
\alias{getngram}
\title{Get a sparse matrix of ngrams from a word, given an existing ngram matrix}
\usage{
getngram(fit, newtxt)
}
\arguments{
\item{fit}{existing ngram counts (vector) to match}

\item{newtxt}{the new word to match ngrams and count}
}
\description{
Given a vector of words and the length of grams to slice, this will
return a matrix of counts each ngram appears in the words.
}

