% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{gram345}
\alias{gram345}
\title{frequency count of domain name based n-grams}
\format{frequency count}
\description{
A dataset containing a transformed frequency count of {3,4,5}-grams extracted
from a dictionary of words.
}
\keyword{datasets}

