\docType{data}
\name{graphs3}
\alias{graphs3}
\title{All Decomposable Graphical Models on Three Lists}
\format{A list of lists. graphs3[[i]] is the \code{i}th model under consideration. This consists of graphs3[[i]]$C, all of the cliques in that model, and graphs3[[i]]$S, the separators.  }
\usage{
graphs3
}
\description{
This dataset contains all of the cliques and separators for each of the decomposable graphical models on three lists. On three lists, this is all of the models. }
\keyword{datasets}