% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphs4}
\alias{graphs4}
\title{All Decomposable Graphical Models on Four Lists}
\format{
A list of lists. graphs4[[i]] is the \code{i}th model under
consideration. This consists of graphs4[[i]]$C, all of the cliques in that
model, and graphs4[[i]]$S, the separators.
}
\usage{
data(graphs4)
}
\description{
This dataset contains all of the cliques and separators for each of the
decomposable graphical models on four lists.
}
\keyword{datasets}
