% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphs5}
\alias{graphs5}
\title{All Decomposable Graphical Models on Five Lists}
\format{
A list of lists. graphs5[[i]] is the \code{i}th model under
consideration. This consists of graphs5[[i]]$C, all of the cliques in that
model, and graphs5[[i]]$S, the separators.
}
\usage{
data(graphs5)
}
\description{
This dataset contains all of the cliques and separators for each of the
decomposable graphical models on five lists.
}
\keyword{datasets}
