% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSummaries.R
\name{htmlSummary}
\alias{htmlSummary}
\title{Save summary to html with plots and graphs}
\usage{
htmlSummary(filepath, weights, N, graphs=NULL, levels=c(0.025, 0.975), nrows = 10, size=60, subplots=TRUE, bg=grDevices::rgb(1,1,1,0), auto=TRUE, ...)
}
\arguments{
\item{filepath}{Filepath (without ".html" file extension) at which to save the html table. Figures are saved in the same folder.}

\item{weights}{Matrix of posterior weights for the population size. Each row should be a (possibly un-normalized) posterior distribution on the parameter `N`, as in the result of `bma.cr`.}

\item{N}{Population sizes under consideration. Each row of `weights` represents a distribution on `N`.}

\item{graphs}{Optional list of decomposable graphs corresponding to the model for each row of `weights`.}

\item{levels}{Posterior quantiles. Default is `c(0.025, 0.975)`.}

\item{nrows}{Posterior summaries are given for the `nrows` models with top posterior probability. Default is 10.}

\item{size}{Pixel height of the svg figures included in the html table. Default is 60.}

\item{subplots}{Set to TRUE to plot posterior distributions in the table, FALSE otherwise. Default is true.}

\item{bg}{Figures background color. Default is `rgb(1,1,1,0)` for a transparent background.}

\item{auto}{Parameter passed to `xtable::xtable`. Set to TRUE for auto-formatting of table entries.}

\item{...}{Other parameters passed to xtable::xtable in order to format the table.}
}
\description{
Save summary to html with plots and graphs
}
