% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dga.R
\name{ngram.name}
\alias{ngram.name}
\title{Calculate the entropy for a given string}
\usage{
ngram.name(instr, n)
}
\arguments{
\item{instr}{character or vector of characters to calculate entropy on}

\item{n}{one or more lengths of n-grams to cut}
}
\description{
Given a string this will return a numric with the entropy of the 
characters in the string.
}

