% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot decomposable graphical model}
\usage{
plotGraph(
  graph,
  labels = TRUE,
  layout = "circle",
  normalize = FALSE,
  edge.color = "black",
  edge.width = 1.5,
  label.font = 2,
  label.norm = "OO",
  bg = grDevices::rgb(1, 1, 1, 0),
  ...
)
}
\arguments{
\item{graph}{decomposable graph as in `data(graphs5)`.}

\item{labels}{Parameter passed to `qgraph::qgraph` with default TRUE.}

\item{layout}{Parameter passed to `qgraph::qgraph` with default "circle".}

\item{normalize}{Parameter passed to `qgraph::qgraph` with default FALSE.}

\item{edge.color}{Parameter passed to `qgraph::qgraph` with default "black".}

\item{edge.width}{Parameter passed to `qgraph::qgraph` with default 1.5.}

\item{label.font}{Parameter passed to `qgraph::qgraph` with default 2.}

\item{label.norm}{Parameter passed to `qgraph::qgraph` with default "OO".}

\item{bg}{Parameter passed to `qgraph::qgraph` with default `rgb(1,1,1,0)` for a transparent background.}

\item{...}{Further arguments passed to `qgraph::qgraph`.}
}
\description{
Plot decomposable graphical model
}
