% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorMode.R
\name{posteriorMode}
\alias{posteriorMode}
\title{Compute posterior mode}
\usage{
posteriorMode(weights, N)
}
\arguments{
\item{weights}{Matrix of posterior weights for the population size. Each row should be a (possibly un-normalized) posterior distribution on the parameter `N`, as in the result of `bma.cr`.}

\item{N}{Population sizes under consideration. Each row of `weights` represents a distribution on `N`.}
}
\value{
vector of computed posterior modes, one for each row of `weights`.
}
\description{
Compute posterior mode
}
\examples{
delta <- .5
Y = c(0,27,37,19,4,4,1,1,97,22,37,25,2,1,3,5,83,36,34,18,3,5,0,2,30,5,23,8,0,3,0,2)
Y <- array(Y, dim=c(2,2,2,2,2))
Nmissing <- 1:300
N <- Nmissing + sum(Y)
data(graphs5)
weights <- bma.cr(Y,  Nmissing, delta, graphs5)

# Full posterior mode
posteriorMode(colSums(weights), N)

# Model-wise posterior modes.
posteriorMode(weights, N)

}
