% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorSummaries.R
\name{posteriorSummaryTable}
\alias{posteriorSummaryTable}
\title{Table summary of posterior distributions}
\usage{
posteriorSummaryTable(weights, N, levels=c(0.025, 0.975), nrows = 10, ndigits=3)
}
\arguments{
\item{weights}{Matrix of posterior weights for the population size. Each row should be a (possibly un-normalized) posterior distribution on the parameter `N`, as in the result of `bma.cr`.}

\item{N}{Population sizes under consideration. Each row of `weights` represents a distribution on `N`.}

\item{levels}{Posterior quantiles. Default is `c(0.025, 0.975)`.}

\item{nrows}{Posterior summaries are given for the `nrows` models with top posterior probability. Default is 10.}

\item{ndigits}{Rounding number of digits for the model-wise posterior probabilities. Default is 3.}
}
\value{
Dataframe with columns `Model` for the model indices (as rows of the `weights` argument), `Prob.` for the posterior probabilities, `Bayes. est` for the Bayes estimators, `Mode` for the posterior and columns corresponding to posterior quantile levels.
}
\description{
This function computes the population size Bayes estimator, posterior mode and posterior quantiles for models with top posterior probability.
}
\examples{
delta <- .5
Y = c(0,27,37,19,4,4,1,1,97,22,37,25,2,1,3,5,83,36,34,18,3,5,0,2,30,5,23,8,0,3,0,2)
Y <- array(Y, dim=c(2,2,2,2,2))
Nmissing <- 1:300
N <- Nmissing + sum(Y)
data(graphs5)
weights <- bma.cr(Y,  Nmissing, delta, graphs5)

# Compute summaries for the top 5 models
posteriorSummaryTable(weights, N, nrows=5)

# Compute summaries for the full posterior distribution
posteriorSummaryTable(colSums(weights), N)

}
