% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wmatch}
\alias{wmatch}
\title{Calculate the proportion of a string that matches dictionary words}
\usage{
wmatch(text, cnum = T)
}
\arguments{
\item{text}{character object (or vector) of words to match}

\item{cnum}{logical whether to count numbers as a match or not, if not counted and the word contains numbers it will never be 100\% match.}
}
\description{
Given a word (or vector of character/words), this will do substring
matches of the string against a dictionary of known words and
return the percentage of letters in the word that are explained
by the dictionary.
}

