% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_functions.R
\name{cfunction}
\alias{cfunction}
\title{A Helper Function for make.strata}
\usage{
cfunction(x, nlist)
}
\arguments{
\item{x}{capture histories, transformed from binary to decimal}

\item{nlist}{the number of lists}
}
\value{
a table of the number of records with each capture history
}
\description{
A helper function used in make.strata to make list overlap counts.
}
\examples{


## The function is currently defined as
cfunction <- function(x, nlist) {
  out <- table(c(x, 0:(2^nlist - 1))) - 1
}
}
\author{
Kristian Lum \email{kl@hrdag.org}
}
\keyword{stratification}
