% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_functions.R
\name{circle}
\alias{circle}
\title{A Helper Function Used in venn3}
\usage{
circle(x, y, r)
}
\arguments{
\item{x}{the x coordinate of the center of the circle.}

\item{y}{the y coordinate of the center of the circle.}

\item{r}{the radius of the circle}
}
\value{
\item{inds}{the x,y coordinates of the periphery of a circle, to be
used in venn3. }
}
\description{
Takes the parameters of a circle and returns points on its perimeter to be
plotted to make circles for a venn diagram.
}
\examples{


plot(dga:::circle(0, 0, 1), type = "l")
}
\author{
Kristian Lum \email{kl@hrdag.org}
}
\keyword{circle}
