\name{dga-package}
\alias{dga-package}
\alias{dga}
\docType{package}
\title{
Capture-Recapture Estimation using Bayesian Model Averaging}
\description{
Performs Bayesian model averaging over all decomposable graphical models, each representing a different  list dependence structure,  with the goal of estimating the number of uncounted elements from the universe of elements that could have been recorded or ``captured".  The code here is an implementation of the model described in Madigan and York (1997).
}
\details{
\tabular{ll}{
Package: \tab dga\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2015-04-16\cr
License: \tab GPL >=2 \cr
}
 }
\author{
James Johndrow \email{james.johndrow@gmail.com}, Kristian Lum \email{kl@hrdag.org}, and Patrick Ball \email{pball@hrdag.org}

Maintainer: Kristian Lum \email{kl@hrdag.org}
}
\references{
Madigan, David, and Jeremy C. York. "Bayesian methods for estimation of the size of a closed population."  Biometrika 84.1 (1997): 19-31.}


\keyword{ capture-recapture }
\keyword{ multiple systems estimation }
\keyword{ Bayesian model averaging }

