% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_functions.R
\name{ellipse}
\alias{ellipse}
\title{A Helper Function Used by Venn4 to Define the Perimeter of an Ellipse}
\usage{
ellipse(x, y, a, b, alpha)
}
\arguments{
\item{x}{the x coordinate of the center of the ellipse.}

\item{y}{the y coordinate of the center of the ellipse.}

\item{a}{the x-direction radius.}

\item{b}{the y-direction radius.}

\item{alpha}{the angle of rotation of the ellipse}
}
\value{
points that define the perimeter of an ellipse.
}
\description{
Draws the ellipses used in venn4.
}
\examples{


plot(dga:::ellipse(0, 0, .5, .2, 1))
}
\author{
Kristian Lum \email{kl@hrdag.org}
}
\keyword{ellipse}
