% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAfunctions.R
\name{integer.base.b}
\alias{integer.base.b}
\title{Base Converter}
\usage{
integer.base.b(x, b = 2)
}
\arguments{
\item{x}{A number.}

\item{b}{The desired base.}
}
\value{
A number in base b.
}
\description{
Takes a decimal number and converts it to base b.
}
\details{
This was harvested from the internet here:
https://stat.ethz.ch/pipermail/r-help/2003-September/038978.html. Posted by
Spencer Graves.
}
\references{
https://stat.ethz.ch/pipermail/r-help/2003-September/038978.html
}
\author{
Spencer Graves
}
\keyword{binary}
\keyword{decimal}
