% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAfunctions.R
\name{MakeCompMatrix}
\alias{MakeCompMatrix}
\title{Component-wise Matrix of Log Marginal Likelihoods}
\usage{
MakeCompMatrix(p, delta, Y, Nmissing)
}
\arguments{
\item{p}{Number of lists}

\item{delta}{Prior hyperparameter of the Dirichlet distribution.}

\item{Y}{The \code{2^k} matrix of list intersection counts.}

\item{Nmissing}{The vector of possible values for the missing cell.}
}
\value{
A matrix of log marginal likelihoods.
}
\description{
Calls CompLogML to create a matrix of number of possible components by
length(Nmissing) log marginal likelihoods. Calculates the log marginal
likehood of each possible marginal table for every value of Nmissing.
}
\author{
James Johndrow \email{james.johndrow@gmail.com} and Kristian Lum
\email{kl@hrdag.org}
}
\keyword{Bayesian}
\keyword{averaging}
\keyword{likelihood}
\keyword{marginal}
\keyword{model}
