% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_functions.R
\name{sfunction}
\alias{sfunction}
\title{A Helper Function for make.strata.}
\usage{
sfunction(x)
}
\arguments{
\item{x}{capture histories, as numbers}
}
\value{
apply(x, 2, sum)
}
\description{
This is the simplest function ever. It's just an apply to sum across
columns.
}
\examples{



## The function is currently defined as
sfunction <- function(x) {
  out <- apply(x, 2, sum)
}
}
\author{
Kristian Lum \email{kl@hrdag.org}
}
\keyword{stratification}
